Revert storing back-link to DecisionTree in every node
Details: Moving the tree invalidates back-links and there are way too many nodes to update them all. Instead reintroduce the back-link in DecisionSubtreeRef and pass it as an argument when querying child nodes. It is still invalidated if the tree is moved, but while nodes are also used as permanent data, DecisionSubtreeRefs are only used to grand temporary access.

Todo: Make sure the tree is not moved while it has pending DecisionSubtreeRefs.