Face reality. For now IfThenElse is here to stay. Subtree switches may greatly reduce the number of conditional jumps, but they have too many issues.
While ifs have two branches, subtree switches have A LOT. Branch prediction may be right for ifs in 50% of the cases, for switches it's essentially 0%. Additional overhead is another issue, first of all in loads of feature values. Even the L2SubtreeSwitch has 50% more loads then IfThenElse. This is so much and with every other level it gets worse exponentially. Compiling leaf subtrees to switch tables is really efficient, but it's still not good enough to regain the ground we lost before.

This is to be continued with new ideas! Cheers